backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
email-validator>=2.1.0
